refactor: 사진 앨범·카메라 선택 바텀시트 및 picker hook 공용화
핀 작성 화면의 앨범·카메라 선택 로직을 rememberPhotoSourcePicker로 추출해
시민해결사·핀 수정 등 후속 화면에서 재사용할 수 있도록 정리한다.
