+++ app/src/main/java/com/issueissyu/fe/ui/components/PhotoAddSourceBottomSheet.kt
@@ -1,0 +1,96 @@
+package com.issueissyu.fe.ui.components
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.PhotoCamera
+import androidx.compose.material.icons.filled.PhotoLibrary
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.ModalBottomSheet
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.unit.dp
+import com.issueissyu.fe.ui.theme.Gray_5
+import com.issueissyu.fe.ui.theme.IssueTypo
+import com.issueissyu.fe.ui.theme.Text as TextColor
+import com.issueissyu.fe.ui.theme.Title
+import com.issueissyu.fe.ui.theme.White
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun PhotoAddSourceBottomSheet(
+    onDismissRequest: () -> Unit,
+    onAlbumClick: () -> Unit,
+    onCameraClick: () -> Unit,
+) {
+    ModalBottomSheet(
+        onDismissRequest = onDismissRequest,
+        containerColor = White,
+        shape = RoundedCornerShape(topStart = 20.dp, topEnd = 20.dp),
+        dragHandle = null,
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(horizontal = 20.dp, vertical = 16.dp),
+            verticalArrangement = Arrangement.spacedBy(8.dp),
+        ) {
+            Text(
+                text = "사진 추가",
+                style = IssueTypo.Bold18.copy(color = Title),
+            )
+            PhotoSourceOptionRow(
+                icon = Icons.Default.PhotoLibrary,
+                label = "앨범에서 선택",
+                onClick = onAlbumClick,
+            )
+            PhotoSourceOptionRow(
+                icon = Icons.Default.PhotoCamera,
+                label = "카메라로 촬영",
+                onClick = onCameraClick,
+            )
+            Spacer(modifier = Modifier.height(8.dp))
+        }
+    }
+}
+
+@Composable
+private fun PhotoSourceOptionRow(
+    icon: ImageVector,
+    label: String,
+    onClick: () -> Unit,
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .clip(RoundedCornerShape(12.dp))
+            .clickable(onClick = onClick)
+            .padding(horizontal = 12.dp, vertical = 14.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
+        Icon(
+            imageVector = icon,
+            contentDescription = null,
+            tint = Gray_5,
+            modifier = Modifier.size(24.dp),
+        )
+        Text(
+            text = label,
+            style = IssueTypo.Regular15.copy(color = TextColor),
+        )
+    }
+}
