+++ app/src/main/java/com/issueissyu/fe/ui/screens/pincreate/PinCreateScreen.kt
@@ -1,11 +1,5 @@
 package com.issueissyu.fe.ui.screens.pincreate
 
-import android.Manifest
-import android.content.pm.PackageManager
-import android.widget.Toast
-import androidx.activity.compose.rememberLauncherForActivityResult
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
@@ -28,36 +22,26 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Close
 import androidx.compose.material.icons.filled.PhotoCamera
-import androidx.compose.material.icons.filled.PhotoLibrary
 import androidx.compose.material.icons.filled.Place
-import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.runtime.SideEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.core.content.ContextCompat
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.issueissyu.fe.core.constants.PinImageUploadConstraints
-import com.issueissyu.fe.core.media.CapturedImageSaver
+import com.issueissyu.fe.core.media.rememberPhotoSourcePicker
 import com.issueissyu.fe.domain.model.pin.PinCategory
 import com.issueissyu.fe.ui.components.CommonButton
 import com.issueissyu.fe.ui.components.CommonTextField
@@ -76,7 +60,6 @@ import com.issueissyu.fe.ui.theme.White
 import androidx.compose.ui.tooling.preview.Preview
 import coil.compose.AsyncImage
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun PinCreateScreen(
     category: PinCategory,
@@ -91,102 +74,14 @@ fun PinCreateScreen(
     viewModel: PinCreateViewModel = hiltViewModel()
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-    val context = LocalContext.current
-    var showPhotoSourceSheet by remember { mutableStateOf(false) }
 
-    val remainingSlots = PinImageUploadConstraints.MAX_COUNT - uiState.imageUris.size
-    val albumPickRequest = PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
-
-    var launchAlbumPicker by remember { mutableStateOf<(() -> Unit)?>(null) }
-
-    if (remainingSlots > 1) {
-        key(remainingSlots) {
-            val imagePickerLauncher = rememberLauncherForActivityResult(
-                contract = ActivityResultContracts.PickMultipleVisualMedia(maxItems = remainingSlots),
-            ) { uris ->
-                viewModel.addImageUris(uris.map { it.toString() })
-            }
-            SideEffect {
-                launchAlbumPicker = {
-                    imagePickerLauncher.launch(albumPickRequest)
-                }
-            }
-        }
-    } else if (remainingSlots == 1) {
-        key("single-album") {
-            val imagePickerLauncher = rememberLauncherForActivityResult(
-                contract = ActivityResultContracts.PickVisualMedia(),
-            ) { uri ->
-                uri?.let { picked -> viewModel.addImageUris(listOf(picked.toString())) }
-            }
-            SideEffect {
-                launchAlbumPicker = {
-                    imagePickerLauncher.launch(albumPickRequest)
-                }
-            }
-        }
-    } else {
-        SideEffect {
-            launchAlbumPicker = null
-        }
-    }
-
-    val cameraLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.TakePicturePreview(),
-    ) { bitmap ->
-        when {
-            bitmap == null -> Unit
-            else -> {
-                val uri = CapturedImageSaver.saveJpegToCache(context, bitmap, "pin-create")
-                if (uri == null) {
-                    Toast.makeText(context, "사진을 저장하지 못했습니다.", Toast.LENGTH_SHORT).show()
-                } else {
-                    viewModel.addImageUris(listOf(uri))
-                }
-            }
-        }
-    }
-
-    val cameraPermissionLauncher = rememberLauncherForActivityResult(
-        contract = ActivityResultContracts.RequestPermission(),
-    ) { granted ->
-        if (granted) {
-            cameraLauncher.launch(null)
-        } else {
-            Toast.makeText(context, "카메라 권한이 필요합니다.", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    val launchCamera = remember(cameraLauncher, cameraPermissionLauncher, context) {
-        {
-            showPhotoSourceSheet = false
-            val hasPermission = ContextCompat.checkSelfPermission(
-                context,
-                Manifest.permission.CAMERA,
-            ) == PackageManager.PERMISSION_GRANTED
-            if (hasPermission) {
-                cameraLauncher.launch(null)
-            } else {
-                cameraPermissionLauncher.launch(Manifest.permission.CAMERA)
-            }
-        }
-    }
-
-    val launchAlbum: () -> Unit = remember {
-        {
-            showPhotoSourceSheet = false
-            launchAlbumPicker?.invoke()
-            Unit
-        }
-    }
-
-    if (showPhotoSourceSheet) {
-        PhotoAddSourceBottomSheet(
-            onDismissRequest = { showPhotoSourceSheet = false },
-            onAlbumClick = launchAlbum,
-            onCameraClick = launchCamera,
-        )
-    }
+    val photoSourcePicker = rememberPhotoSourcePicker(
+        currentCount = uiState.imageUris.size,
+        maxCount = PinImageUploadConstraints.MAX_COUNT,
+        onImagesPicked = viewModel::addImageUris,
+        cameraFileNamePrefix = "pin-create",
+    )
+    photoSourcePicker.PhotoSourceBottomSheet()
 
     LaunchedEffect(category, pinLat, pinLng, userLat, userLng, address) {
         viewModel.initialize(category, pinLat, pinLng, userLat, userLng, address)
@@ -205,7 +100,7 @@ fun PinCreateScreen(
         onTitleChange = viewModel::onTitleChange,
         onDescriptionChange = viewModel::onDescriptionChange,
         onToneChange = viewModel::onToneChange,
-        onPhotoAddClick = { showPhotoSourceSheet = true },
+        onPhotoAddClick = photoSourcePicker.showSourceSheet,
         onPhotoRemoveClick = viewModel::removeImageUri,
         onSetMainImageClick = viewModel::setMainImageUri,
         onAiDraftClick = viewModel::createAiDraft,
@@ -389,72 +284,6 @@ private fun PhotoUploadSection(
                 )
             }
         }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-private fun PhotoAddSourceBottomSheet(
-    onDismissRequest: () -> Unit,
-    onAlbumClick: () -> Unit,
-    onCameraClick: () -> Unit,
-) {
-    ModalBottomSheet(
-        onDismissRequest = onDismissRequest,
-        containerColor = White,
-        shape = RoundedCornerShape(topStart = 20.dp, topEnd = 20.dp),
-        dragHandle = null,
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(horizontal = 20.dp, vertical = 16.dp),
-            verticalArrangement = Arrangement.spacedBy(8.dp),
-        ) {
-            Text(
-                text = "사진 추가",
-                style = IssueTypo.Bold18.copy(color = Title),
-            )
-            PhotoSourceOptionRow(
-                icon = Icons.Default.PhotoLibrary,
-                label = "앨범에서 선택",
-                onClick = onAlbumClick,
-            )
-            PhotoSourceOptionRow(
-                icon = Icons.Default.PhotoCamera,
-                label = "카메라로 촬영",
-                onClick = onCameraClick,
-            )
-            Spacer(modifier = Modifier.height(8.dp))
-        }
-    }
-}
-
-@Composable
-private fun PhotoSourceOptionRow(
-    icon: androidx.compose.ui.graphics.vector.ImageVector,
-    label: String,
-    onClick: () -> Unit,
-) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(RoundedCornerShape(12.dp))
-            .clickable(onClick = onClick)
-            .padding(horizontal = 12.dp, vertical = 14.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(12.dp),
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = null,
-            tint = Gray_5,
-            modifier = Modifier.size(24.dp),
-        )
-        Text(
-            text = label,
-            style = IssueTypo.Regular15.copy(color = TextColor),
-        )
     }
 }
